requirements.txt
Django==5.1.8
djangorestframework==3.16.0
Pillow==11.2.1
requests==2.32.3
psycopg2-binary==2.9.10
drf-spectacular==0.28.0
djangorestframework-simplejwt==5.5.0